WelcomeView -> saveWelcomeView
### What's done:
 * Refactored WelcomeView - now it is FC instead of ClassComponent
 * Fixed frontend compiler warnings
