+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/basic/organizations/OrganizationToolsMenu.kt
@@ -125,7 +125,7 @@ private fun organizationToolsMenu() = FC<OrganizationToolsMenuProps> { props ->
                 column(id = "description", header = "Description") {
                     Fragment.create {
                         td {
-                            +(it.value.description ?: "Description not provided")
+                            +it.value.description
                         }
                     }
                 }
