+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/usersettings/UserSettingsProfileMenuView.kt
@@ -3,7 +3,7 @@ package com.saveourtool.save.frontend.components.views.usersettings
 import com.saveourtool.save.frontend.components.basic.cardComponent
 import com.saveourtool.save.frontend.components.inputform.InputTypes
 
-import react.FC
+import react.VFC
 import react.dom.html.ReactHTML.button
 import react.dom.html.ReactHTML.div
 import react.dom.html.ReactHTML.hr
@@ -16,7 +16,7 @@ import web.html.InputType
 class UserSettingsProfileMenuView : UserSettingsView() {
     private val card = cardComponent(isBordered = false, hasBg = true)
     @Suppress("TOO_LONG_FUNCTION", "LongMethod", "EMPTY_BLOCK_STRUCTURE_ERROR")
-    override fun renderMenu(): FC<UserSettingsProps> = FC { props ->
+    override fun renderMenu(): VFC = VFC {
         card {
             div {
                 className = ClassName("row mt-2 ml-2 mr-2")
