+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/utils/Particles.kt
@@ -11,9 +11,12 @@ import react.react
 /**
  * @param enabled true if particles enabled
  */
-fun ChildrenBuilder.particles(enabled: Boolean = true) {
+fun ChildrenBuilder.particles(
+    enabled: Boolean = true,
+) {
     if (enabled) {
-        // FixMe: Note that they block user interactions. Particles are superimposed on top of the view in some transitions
+        // FixMe: Note that they block user interactions.
+        // Particles are superimposed on top of the view in some transitions
         // https://github.com/matteobruni/tsparticles/discussions/4489
         Particles::class.react {
             id = "tsparticles"
